Move input streams to common
