+++ CMakeLists.txt
@@ -41,11 +41,11 @@ option(VELOX_ENABLE_EXAMPLES
        "Build examples. This will enable VELOX_ENABLE_EXPRESSION automatically."
        OFF)
 option(VELOX_ENABLE_SUBSTRAIT "Buid Substrait-to-Velox converter." OFF)
-option(VELOX_ENABLE_BENCHMARKS "Build velox top level benchmarks." OFF)
-option(VELOX_ENABLE_BENCHMARKS_BASIC "Build velox basic benchmarks." OFF)
+option(VELOX_ENABLE_BENCHMARKS "Build velox top level benchmarks." ON)
+option(VELOX_ENABLE_BENCHMARKS_BASIC "Build velox basic benchmarks." ON)
 option(VELOX_ENABLE_S3 "Build S3 Connector" OFF)
-option(VELOX_ENABLE_PARQUET "Enable Parquet support" OFF)
-option(VELOX_BUILD_TEST_UTILS "Enable Velox test utilities" OFF)
+option(VELOX_ENABLE_PARQUET "Enable Parquet support" ON)
+option(VELOX_BUILD_TEST_UTILS "Enable Velox test utilities" ON)
 
 if(${VELOX_BUILD_MINIMAL})
   # Enable and disable components for velox base build
