+++ velox/dwio/common/io/BufferedOutputStream.cpp
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-#include "velox/dwio/dwrf/common/OutputStream.h"
+#include "BufferedOutputStream.h"
 #include "velox/dwio/common/exception/Exception.h"
 
-namespace facebook::velox::dwrf {
+namespace facebook::velox::dwio::common::io {
 
 bool BufferedOutputStream::Next(
     void** buffer,
@@ -74,4 +74,4 @@ uint64_t AppendOnlyBufferedStream::flush() {
   return outStream_->flush();
 }
 
-} // namespace facebook::velox::dwrf
+} // namespace facebook::velox::dwio::common::io {
