+++ velox/dwio/common/io/CMakeLists.txt
@@ -11,31 +11,16 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-add_subdirectory(compression)
-add_subdirectory(encryption)
-add_subdirectory(exception)
-if(${VELOX_BUILD_TESTING})
-  add_subdirectory(tests)
-endif()
-add_library(
-  velox_dwio_common
-  ColumnSelector.cpp
-  DataBuffer.cpp
-  DataSink.cpp
-  DwioMetricsLog.cpp
-  IoStatistics.cpp
-  MemoryInputStream.cpp
-  Options.cpp
-  ReaderFactory.cpp
-  ScanSpec.cpp
-  TypeWithId.cpp)
 
-target_link_libraries(
-  velox_dwio_common
-  velox_exception
-  velox_dwio_common_encryption
-  velox_dwio_common_compression
-  velox_memory
-  Boost::regex
-  ${FOLLY_WITH_DEPENDENCIES}
-  glog::glog)
+add_library(
+        velox_dwio_common_io
+        InputStream.cpp
+        BufferedInput.cpp
+        CachedBufferedInput.cpp
+        SeekableInputStream.cpp
+        CacheInputStream.cpp
+        PagedInputStream.cpp
+        BufferedOutputStream.cpp
+    )
+
+target_link_libraries(velox_dwio_common_input ${FOLLY_WITH_DEPENDENCIES})
