+++ velox/dwio/common/io/CachedBufferedInput.h
@@ -16,18 +16,18 @@
 
 #pragma once
 
+#include "CacheInputStream.h"
+#include "InputStream.h"
 #include "velox/common/caching/FileGroupStats.h"
 #include "velox/common/caching/ScanTracker.h"
 #include "velox/common/caching/SsdCache.h"
-#include "velox/dwio/common/InputStream.h"
-#include "velox/dwio/dwrf/common/BufferedInput.h"
-#include "velox/dwio/dwrf/common/CacheInputStream.h"
+#include "velox/dwio/common/io/BufferedInput.h"
 
 #include <folly/Executor.h>
 
 DECLARE_int32(cache_load_quantum);
 
-namespace facebook::velox::dwrf {
+namespace facebook::velox::dwio::common::io {
 
 // Abstract class for owning an InputStream and related structures
 // like pins into file handle caches. TODO: Make file handle cache
@@ -238,4 +238,4 @@ class CachedBufferedInputFactory : public BufferedInputFactory {
   int32_t loadQuantum_;
   int32_t maxCoalesceDistance_;
 };
-} // namespace facebook::velox::dwrf
+} // namespace facebook::velox::dwio::common::io
