+++ velox/dwio/common/io/MemoryInputStream.cpp
@@ -18,10 +18,7 @@
 
 #include <cstring>
 
-namespace facebook {
-namespace velox {
-namespace dwio {
-namespace common {
+namespace facebook::velox::dwio::common::io {
 
 uint64_t MemoryInputStream::getLength() const {
   return size;
@@ -58,7 +55,4 @@ const void* MemoryInputStream::readReferenceOnly(
   return buffer + offset;
 }
 
-} // namespace common
-} // namespace dwio
-} // namespace velox
-} // namespace facebook
+} // namespace facebook::velox::dwio::common::io {
